rs6000-protos.h (rs6000_builtin_is_supported_p): New prototype.
gcc/ChangeLog:

2018-04-13  Kelvin Nilsen  <kelvin@gcc.gnu.org>

	* config/rs6000/rs6000-protos.h (rs6000_builtin_is_supported_p):
	New prototype.
	* config/rs6000/rs6000-c.c (altivec_resolve_overloaded_builtin):
	Add note to error message to explain internal mapping of overloaded
	built-in function name to non-overloaded built-in function name.
	* config/rs6000/rs6000.c (rs6000_builtin_is_supported_p): New
	function.

gcc/testsuite/ChangeLog:

2018-04-13  Kelvin Nilsen  <kelvin@gcc.gnu.org>

	* gcc.target/powerpc/bfp/scalar-extract-sig-5.c: Simplify to
	prevent cascading of errors and change expected error message.
	* gcc.target/powerpc/bfp/scalar-test-neg-4.c: Restrict this test
	to 64-bit targets.
	* gcc.target/powerpc/bfp/scalar-test-data-class-8.c: Likewise.
	* gcc.target/powerpc/bfp/scalar-test-data-class-9.c: Likewise.
	* gcc.target/powerpc/bfp/scalar-test-data-class-10.c: Likewise.
	* gcc.target/powerpc/bfp/scalar-insert-exp-11.c: Change expected
	error message.
	* gcc.target/powerpc/bfp/scalar-extract-exp-5.c: Likewise.

From-SVN: r259442

+++ gcc/testsuite/gcc.target/powerpc/bfp/scalar-extract-sig-5.c
@@ -8,10 +8,10 @@
    error because the builtin requires 64 bits.  */
 #include <altivec.h>
 
-unsigned __int128 /* { dg-error "'__int128' is not supported on this target" } */
+unsigned long long int
 get_significand (__ieee128 *p)
 {
   __ieee128 source = *p;
 
-  return __builtin_vec_scalar_extract_sig (source); /* { dg-error "builtin function '__builtin_vec_scalar_extract_sig' not supported in this compiler configuration" } */
+  return (long long int) __builtin_vec_scalar_extract_sig (source); /* { dg-error "requires ISA 3.0 IEEE 128-bit floating point" } */
 }
